// login/src/shared/enums/permission.ts
export enum Permission {
    MANAGE_GROUP = 'MANAGE_GROUP',
    MANAGE_MEMBERS = 'MANAGE_MEMBERS',
    PROMOTE_MEMBERS = 'PROMOTE_MEMBERS',
    DEMOTE_MEMBERS = 'DEMOTE_MEMBERS',
    CREATE_TEAMS = 'CREATE_TEAMS',
    DELETE_TEAMS = 'DELETE_TEAMS',
    CREATE_CONTENT = 'CREATE_CONTENT',
    DELETE_CONTENT = 'DELETE_CONTENT',
    INVITE_MEMBERS = 'INVITE_MEMBERS',
    ACCEPT_MEMBERS = 'ACCEPT_MEMBERS',
    REMOVE_MEMBERS = 'REMOVE_MEMBERS',
    EDIT_CONTENT = 'EDIT_CONTENT',
    JOIN_CONTENT = 'JOIN_CONTENT'
}